move `white-card` to `shared` folder
