+++ packages/hash/frontend/src/pages/[account-slug]/entities/[entity-id].page/entity-editor/types-section.tsx
@@ -2,7 +2,7 @@ import { faAsterisk } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@hashintel/hash-design-system/fontawesome-icon";
 import { Box, Typography } from "@mui/material";
 import { getPersistedEntityType } from "../../../../../lib/subgraph";
-import { WhiteCard } from "../../../types/entity-type/white-card";
+import { WhiteCard } from "../../../shared/white-card";
 import { useEntityEditor } from "./entity-editor-context";
 import { EntitySection } from "./shared/entity-section";
