+++ packages/hash/frontend/src/pages/[account-slug]/shared/white-card.tsx
